requirements.txt
faster-whisper==0.10.0
ffmpeg-python
flask
gunicorn